requirements.txt
streamlit>=1.36
pypdf>=4.2
python-docx>=1.1
docx2txt>=0.8
tiktoken>=0.7
numpy>=1.26
scipy>=1.12
scikit-learn>=1.5
openai>=1.40
google-generativeai>=0.7
python-dotenv>=1.0
reportlab>=4.2
